refactor(dependency-freshness): consolidate duplicated integration tests
Move workspace tests from per-PM files (cargo.rs, npm.rs, pnpm.rs) into
parameterized tests in package_managers.rs. Parameterize pnpm unit tests
that shared similar structure (skip conditions, location resolution).
